backend/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
google
genai
generativeai
google-generativeai==0.8.3
faster-whisper==1.0.3
#ffmpeg-python==0.2.0
Werkzeug==3.0.4
flask-cors